Initiated process for breaking out classes into separate files and changing naming conventions to match C# standards.
