+++ Plivo/packages.config
@@ -1,6 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="RestSharp" version="104.1" targetFramework="net40" />
-  <package id="RestSharp" version="104.2.0" targetFramework="net40" />
   <package id="RestSharp" version="104.4.0" targetFramework="net40" />
 </packages>
